refactor: split server.go into focused modules with matching test files
Split the 770-line server.go into four focused files:

  server.go     (256 lines) — Server struct, Config, constructor, Start/Stop, accessors
  handlers.go   (204 lines) — System HTTP handlers (liveness, health, status, version, metrics, admin)
  helpers.go    (189 lines) — JSON response helpers, error handling, pagination, query params
  middleware.go (151 lines) — Recovery, logging, content-type, CORS, auth middleware setup

Split server_test.go to match:

  server_test.go     (420 lines) — Lifecycle, config, state, concurrency tests
  handlers_test.go   (129 lines) — Built-in handler endpoint tests
  helpers_test.go    (318 lines) — JSON/pagination/error/edge-case/benchmark tests
  middleware_test.go (120 lines) — Recovery, content-type, responseWriter tests

routes.go (131 lines) was already extracted and remains unchanged.

No behavioral changes — pure mechanical split. All existing tests pass.
